chore: moved utility js since it was just there
