+++ src/pages/request/submit/request_page.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { collection, setDoc, getDoc, doc } from 'firebase/firestore'
 import { auth, db, storage } from '../../../firebase_config.js'
-import { get_date } from '../../../js/utility.js'
+import { get_date } from '../../../utility.js'
 import { WARD_API, PLACEHOLDER_IMAGE } from '../../../constants.js'
 import Navbar from '../../../components/nav_bar/nav_bar.js'
 import RequestForm from '../../../components/request_form/request_form.js'
